refactor(jobs): move MLJob serializers from schemas.py to serializers.py
Follows the codebase convention: schemas.py holds OpenApiParameter
definitions and Pydantic models, serializers.py holds DRF serializers.
